fiemap: Start using new callback from fiemap_ctx
Now that fiemap methods are updated to work based on the new fiemap_ctx
structure, update the code to use the new callback.
The current fiemap_fill_next_extent will be used for user calls of
->fiemap methods, as its done today, but passed in as a callback to
fiemap_ctx. So, rename it to ifemap_fill_usr_extent().

The 'new' fiemap_fill_next_extent() will now be just a helper to call
fiemap_ctx callback.