fiemap: Use a callback to fill fiemap extents
As a goal to enable fiemap infrastructure to be used by fibmap too, we need a
way to use different helpers to fill extent data, depending on its usage. One
helper to fill extent data stored in user address space (used in fiemap), and
another fo fill extent data stored in kernel address space (will be used in
fibmap).

This patch sets up the usage of a callback to be used to fill in the extents.
It transforms the current fiemap_fill_next_extent, into a simple helper to call
the callback, avoiding unneeded changes on any filesystem, and reutilizes the
original function as the callback used by FIEMAP.

V3:
	- Rebase to current linux-next
	- Fix conflict on rebase
	- Merge this patch into patch 07 from V2
	- Rename fi_extents_start to fi_cb_data

V2:
	- Now based on the rework on fiemap_extent_info (previous was
	  based on fiemap_ctx)